Fixed the call to ctypes's find_library function.
The name parameter of the find_library function is the library name
without any prefix like lib, suffix like .so, .dylib or version number.
Since the name parameter was 'libocsync', it didn't find the library,
but with 'ocsync' as parameter it did find it.

I wouldn't be surprised if the hardcoded paths can now be removed, so
only relying on the find_library function.